LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/animations/background-size-interpolation-expected.txt


PASS CSS Transitions: property <background-size> from neutral to [20px 20px, 0px 0px] at (-0.25) should be [ 7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px]
PASS CSS Transitions: property <background-size> from neutral to [20px 20px, 0px 0px] at (0) should be [10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px]
PASS CSS Transitions: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.25) should be [12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px]
PASS CSS Transitions: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.5) should be [15.0px 15.0px,  5.0px  5.0px, 15.0px 15.0px,  5.0px  5.0px]
PASS CSS Transitions: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.75) should be [17.5px 17.5px,  2.5px  2.5px, 17.5px 17.5px,  2.5px  2.5px]
PASS CSS Transitions: property <background-size> from neutral to [20px 20px, 0px 0px] at (1) should be [20.0px 20.0px,  0.0px  0.0px, 20.0px 20.0px,  0.0px  0.0px]
FAIL CSS Transitions: property <background-size> from neutral to [20px 20px, 0px 0px] at (1.25) should be [22.5px 22.5px,  0.0px  0.0px, 22.5px 22.5px,  0.0px  0.0px] assert_equals: expected "22.5px 22.5px , 0px 0px , 22.5px 22.5px , 0px 0px " but got "22.5px 22.5px , - 2.5px - 2.5px , 22.5px 22.5px , - 2.5px - 2.5px "
PASS CSS Transitions with transition: all: property <background-size> from neutral to [20px 20px, 0px 0px] at (-0.25) should be [ 7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px]
PASS CSS Transitions with transition: all: property <background-size> from neutral to [20px 20px, 0px 0px] at (0) should be [10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px]
PASS CSS Transitions with transition: all: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.25) should be [12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px]
PASS CSS Transitions with transition: all: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.5) should be [15.0px 15.0px,  5.0px  5.0px, 15.0px 15.0px,  5.0px  5.0px]
PASS CSS Transitions with transition: all: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.75) should be [17.5px 17.5px,  2.5px  2.5px, 17.5px 17.5px,  2.5px  2.5px]
PASS CSS Transitions with transition: all: property <background-size> from neutral to [20px 20px, 0px 0px] at (1) should be [20.0px 20.0px,  0.0px  0.0px, 20.0px 20.0px,  0.0px  0.0px]
FAIL CSS Transitions with transition: all: property <background-size> from neutral to [20px 20px, 0px 0px] at (1.25) should be [22.5px 22.5px,  0.0px  0.0px, 22.5px 22.5px,  0.0px  0.0px] assert_equals: expected "22.5px 22.5px , 0px 0px , 22.5px 22.5px , 0px 0px " but got "22.5px 22.5px , - 2.5px - 2.5px , 22.5px 22.5px , - 2.5px - 2.5px "
PASS CSS Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (-0.25) should be [ 7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px]
PASS CSS Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0) should be [10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px]
PASS CSS Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.25) should be [12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px]
PASS CSS Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.5) should be [15.0px 15.0px,  5.0px  5.0px, 15.0px 15.0px,  5.0px  5.0px]
PASS CSS Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.75) should be [17.5px 17.5px,  2.5px  2.5px, 17.5px 17.5px,  2.5px  2.5px]
PASS CSS Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (1) should be [20.0px 20.0px,  0.0px  0.0px, 20.0px 20.0px,  0.0px  0.0px]
FAIL CSS Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (1.25) should be [22.5px 22.5px,  0.0px  0.0px, 22.5px 22.5px,  0.0px  0.0px] assert_equals: expected "22.5px 22.5px , 0px 0px , 22.5px 22.5px , 0px 0px " but got "22.5px 22.5px , - 2.5px - 2.5px , 22.5px 22.5px , - 2.5px - 2.5px "
PASS Web Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (-0.25) should be [ 7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px]
FAIL Web Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0) should be [10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px, 10.0px 10.0px] assert_equals: expected "10px 10px , 10px 10px , 10px 10px , 10px 10px " but got "7.5px 7.5px , 12.5px 12.5px , 7.5px 7.5px , 12.5px 12.5px "
FAIL Web Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.25) should be [12.5px 12.5px,  7.5px  7.5px, 12.5px 12.5px,  7.5px  7.5px] assert_equals: expected "12.5px 12.5px , 7.5px 7.5px , 12.5px 12.5px , 7.5px 7.5px " but got "10.63px 10.63px , 9.38px 9.38px , 10.63px 10.63px , 9.38px 9.38px "
FAIL Web Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.5) should be [15.0px 15.0px,  5.0px  5.0px, 15.0px 15.0px,  5.0px  5.0px] assert_equals: expected "15px 15px , 5px 5px , 15px 15px , 5px 5px " but got "15.31px 15.31px , 4.69px 4.69px , 15.31px 15.31px , 4.69px 4.69px "
FAIL Web Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (0.75) should be [17.5px 17.5px,  2.5px  2.5px, 17.5px 17.5px,  2.5px  2.5px] assert_equals: expected "17.5px 17.5px , 2.5px 2.5px , 17.5px 17.5px , 2.5px 2.5px " but got "18.83px 18.83px , 1.17px 1.17px , 18.83px 18.83px , 1.17px 1.17px "
PASS Web Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (1) should be [20.0px 20.0px,  0.0px  0.0px, 20.0px 20.0px,  0.0px  0.0px]
FAIL Web Animations: property <background-size> from neutral to [20px 20px, 0px 0px] at (1.25) should be [22.5px 22.5px,  0.0px  0.0px, 22.5px 22.5px,  0.0px  0.0px] assert_equals: expected "22.5px 22.5px , 0px 0px , 22.5px 22.5px , 0px 0px " but got "20px 20px , 0px 0px , 20px 20px , 0px 0px "
FAIL CSS Transitions: property <background-size> from [initial] to [20px 20px, 0px 0px] at (-0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
PASS CSS Transitions: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS CSS Transitions: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS CSS Transitions: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS CSS Transitions: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
FAIL CSS Transitions with transition: all: property <background-size> from [initial] to [20px 20px, 0px 0px] at (-0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions with transition: all: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions with transition: all: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
PASS CSS Transitions with transition: all: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS CSS Transitions with transition: all: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS CSS Transitions with transition: all: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS CSS Transitions with transition: all: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (-0.3) should be [initial]
PASS CSS Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0) should be [initial]
PASS CSS Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.3) should be [initial]
PASS CSS Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (-0.3) should be [initial]
PASS Web Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0) should be [initial]
PASS Web Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.3) should be [initial]
PASS Web Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [initial] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
PASS CSS Transitions: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (-0.25) should be [120px 120px, 125px 125px, 120px 120px, 125px 125px]
PASS CSS Transitions: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px]
PASS CSS Transitions: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.25) should be [ 80px  80px,  75px  75px,  80px  80px,  75px  75px]
PASS CSS Transitions: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.5) should be [ 60px  60px,  50px  50px,  60px  60px,  50px  50px]
PASS CSS Transitions: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.75) should be [ 40px  40px,  25px  25px,  40px  40px,  25px  25px]
PASS CSS Transitions: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1) should be [ 20px  20px,   0px   0px,  20px  20px,   0px   0px]
FAIL CSS Transitions: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "0px 0px , - 25px - 25px , 0px 0px , - 25px - 25px "
PASS CSS Transitions with transition: all: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (-0.25) should be [120px 120px, 125px 125px, 120px 120px, 125px 125px]
PASS CSS Transitions with transition: all: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px]
PASS CSS Transitions with transition: all: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.25) should be [ 80px  80px,  75px  75px,  80px  80px,  75px  75px]
PASS CSS Transitions with transition: all: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.5) should be [ 60px  60px,  50px  50px,  60px  60px,  50px  50px]
PASS CSS Transitions with transition: all: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.75) should be [ 40px  40px,  25px  25px,  40px  40px,  25px  25px]
PASS CSS Transitions with transition: all: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1) should be [ 20px  20px,   0px   0px,  20px  20px,   0px   0px]
FAIL CSS Transitions with transition: all: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "0px 0px , - 25px - 25px , 0px 0px , - 25px - 25px "
PASS CSS Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (-0.25) should be [120px 120px, 125px 125px, 120px 120px, 125px 125px]
PASS CSS Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px]
PASS CSS Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.25) should be [ 80px  80px,  75px  75px,  80px  80px,  75px  75px]
PASS CSS Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.5) should be [ 60px  60px,  50px  50px,  60px  60px,  50px  50px]
PASS CSS Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.75) should be [ 40px  40px,  25px  25px,  40px  40px,  25px  25px]
PASS CSS Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1) should be [ 20px  20px,   0px   0px,  20px  20px,   0px   0px]
FAIL CSS Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "0px 0px , - 25px - 25px , 0px 0px , - 25px - 25px "
FAIL Web Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (-0.25) should be [120px 120px, 125px 125px, 120px 120px, 125px 125px] assert_equals: expected "120px 120px , 125px 125px , 120px 120px , 125px 125px " but got "7.5px 7.5px , 12.5px 12.5px , 7.5px 7.5px , 12.5px 12.5px "
FAIL Web Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px] assert_equals: expected "100px 100px , 100px 100px , 100px 100px , 100px 100px " but got "7.5px 7.5px , 12.5px 12.5px , 7.5px 7.5px , 12.5px 12.5px "
FAIL Web Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.25) should be [ 80px  80px,  75px  75px,  80px  80px,  75px  75px] assert_equals: expected "80px 80px , 75px 75px , 80px 80px , 75px 75px " but got "10.63px 10.63px , 9.38px 9.38px , 10.63px 10.63px , 9.38px 9.38px "
FAIL Web Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.5) should be [ 60px  60px,  50px  50px,  60px  60px,  50px  50px] assert_equals: expected "60px 60px , 50px 50px , 60px 60px , 50px 50px " but got "15.31px 15.31px , 4.69px 4.69px , 15.31px 15.31px , 4.69px 4.69px "
FAIL Web Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (0.75) should be [ 40px  40px,  25px  25px,  40px  40px,  25px  25px] assert_equals: expected "40px 40px , 25px 25px , 40px 40px , 25px 25px " but got "18.83px 18.83px , 1.17px 1.17px , 18.83px 18.83px , 1.17px 1.17px "
PASS Web Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1) should be [ 20px  20px,   0px   0px,  20px  20px,   0px   0px]
FAIL Web Animations: property <background-size> from [inherit] to [20px 20px, 0px 0px] at (1.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "20px 20px , 0px 0px , 20px 20px , 0px 0px "
FAIL CSS Transitions: property <background-size> from [unset] to [20px 20px, 0px 0px] at (-0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
PASS CSS Transitions: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS CSS Transitions: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS CSS Transitions: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS CSS Transitions: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
FAIL CSS Transitions with transition: all: property <background-size> from [unset] to [20px 20px, 0px 0px] at (-0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions with transition: all: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
FAIL CSS Transitions with transition: all: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.3) should be [20px 20px, 0px 0px] assert_equals: expected "20px 20px , 0px 0px , 20px 20px , 0px 0px " but got "auto , auto , auto , auto "
PASS CSS Transitions with transition: all: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS CSS Transitions with transition: all: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS CSS Transitions with transition: all: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS CSS Transitions with transition: all: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (-0.3) should be [unset]
PASS CSS Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0) should be [unset]
PASS CSS Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.3) should be [unset]
PASS CSS Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS CSS Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (-0.3) should be [unset]
PASS Web Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0) should be [unset]
PASS Web Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.3) should be [unset]
PASS Web Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.5) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (0.6) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1) should be [20px 20px, 0px 0px]
PASS Web Animations: property <background-size> from [unset] to [20px 20px, 0px 0px] at (1.5) should be [20px 20px, 0px 0px]
FAIL CSS Transitions: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (-0.25) should be [ 0px auto,  0px  0px, contain, cover] assert_equals: expected "0px , 0px 0px , contain , cover " but got "- 10px , - 10px - 10px , contain , cover "
PASS CSS Transitions: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0) should be [ 0px auto,  0px  0px, contain, cover]
PASS CSS Transitions: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.25) should be [10px auto, 10px 10px, contain, cover]
PASS CSS Transitions: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.5) should be [20px auto, 20px 20px, contain, cover]
PASS CSS Transitions: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.75) should be [30px auto, 30px 30px, contain, cover]
PASS CSS Transitions: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1) should be [40px auto, 40px 40px, contain, cover]
PASS CSS Transitions: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1.25) should be [50px auto, 50px 50px, contain, cover]
FAIL CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (-0.25) should be [ 0px auto,  0px  0px, contain, cover] assert_equals: expected "0px , 0px 0px , contain , cover " but got "- 10px , - 10px - 10px , contain , cover "
PASS CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0) should be [ 0px auto,  0px  0px, contain, cover]
PASS CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.25) should be [10px auto, 10px 10px, contain, cover]
PASS CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.5) should be [20px auto, 20px 20px, contain, cover]
PASS CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.75) should be [30px auto, 30px 30px, contain, cover]
PASS CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1) should be [40px auto, 40px 40px, contain, cover]
PASS CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1.25) should be [50px auto, 50px 50px, contain, cover]
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (-0.25) should be [ 0px auto,  0px  0px, contain, cover] assert_equals: expected "0px , 0px 0px , contain , cover " but got "- 10px , - 10px - 10px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0) should be [ 0px auto,  0px  0px, contain, cover] assert_equals: expected "0px , 0px 0px , contain , cover " but got "0px , 0px 0px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.25) should be [10px auto, 10px 10px, contain, cover] assert_equals: expected "10px , 10px 10px , contain , cover " but got "10px , 10px 10px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.5) should be [20px auto, 20px 20px, contain, cover] assert_equals: expected "20px , 20px 20px , contain , cover " but got "20px , 20px 20px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.75) should be [30px auto, 30px 30px, contain, cover] assert_equals: expected "30px , 30px 30px , contain , cover " but got "30px , 30px 30px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1) should be [40px auto, 40px 40px, contain, cover] assert_equals: expected "40px , 40px 40px , contain , cover " but got "40px , 40px 40px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1.25) should be [50px auto, 50px 50px, contain, cover] assert_equals: expected "50px , 50px 50px , contain , cover " but got "50px , 50px 50px , auto , auto "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (-0.25) should be [ 0px auto,  0px  0px, contain, cover] assert_equals: expected "0px , 0px 0px , contain , cover " but got "- 10px , - 10px - 10px , auto , auto "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0) should be [ 0px auto,  0px  0px, contain, cover] assert_equals: expected "0px , 0px 0px , contain , cover " but got "0px , 0px 0px , auto , auto "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.25) should be [10px auto, 10px 10px, contain, cover] assert_equals: expected "10px , 10px 10px , contain , cover " but got "10px , 10px 10px , auto , auto "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.5) should be [20px auto, 20px 20px, contain, cover] assert_equals: expected "20px , 20px 20px , contain , cover " but got "20px , 20px 20px , auto , auto "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (0.75) should be [30px auto, 30px 30px, contain, cover] assert_equals: expected "30px , 30px 30px , contain , cover " but got "30px , 30px 30px , auto , auto "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1) should be [40px auto, 40px 40px, contain, cover] assert_equals: expected "40px , 40px 40px , contain , cover " but got "40px , 40px 40px , auto , auto "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px, contain, cover] to [40px auto, 40px 40px, contain, cover] at (1.25) should be [50px auto, 50px 50px, contain, cover] assert_equals: expected "50px , 50px 50px , contain , cover " but got "50px , 50px 50px , auto , auto "
FAIL CSS Transitions: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (-0.3) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "- 12px - 12px , - 12px - 12px , cover , contain "
FAIL CSS Transitions: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "0px 0px , 0px 0px , cover , contain "
FAIL CSS Transitions: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.3) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "12px 12px , 12px 12px , cover , contain "
FAIL CSS Transitions: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "20px 20px , 20px 20px , cover , contain "
FAIL CSS Transitions: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.6) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "24px 24px , 24px 24px , cover , contain "
PASS CSS Transitions: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1) should be [40px 40px, 40px 40px, cover, contain]
FAIL CSS Transitions: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "60px 60px , 60px 60px , cover , contain "
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (-0.3) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "- 12px - 12px , - 12px - 12px , cover , contain "
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "0px 0px , 0px 0px , cover , contain "
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.3) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "12px 12px , 12px 12px , cover , contain "
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "20px 20px , 20px 20px , cover , contain "
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.6) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "24px 24px , 24px 24px , cover , contain "
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1) should be [40px 40px, 40px 40px, cover, contain]
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "60px 60px , 60px 60px , cover , contain "
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (-0.3) should be [0px 0px, 0px 0px, contain, cover] assert_equals: expected "0px 0px , 0px 0px , contain , cover " but got "- 12px - 12px , - 12px - 12px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0) should be [0px 0px, 0px 0px, contain, cover] assert_equals: expected "0px 0px , 0px 0px , contain , cover " but got "0px 0px , 0px 0px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.3) should be [0px 0px, 0px 0px, contain, cover] assert_equals: expected "0px 0px , 0px 0px , contain , cover " but got "12px 12px , 12px 12px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "20px 20px , 20px 20px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.6) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "24px 24px , 24px 24px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "40px 40px , 40px 40px , auto , auto "
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "60px 60px , 60px 60px , auto , auto "
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (-0.3) should be [0px 0px, 0px 0px, contain, cover] assert_equals: expected "0px 0px , 0px 0px , contain , cover " but got "- 12px - 12px , - 12px - 12px , auto , auto "
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0) should be [0px 0px, 0px 0px, contain, cover] assert_equals: expected "0px 0px , 0px 0px , contain , cover " but got "0px 0px , 0px 0px , auto , auto "
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.3) should be [0px 0px, 0px 0px, contain, cover] assert_equals: expected "0px 0px , 0px 0px , contain , cover " but got "12px 12px , 12px 12px , auto , auto "
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "20px 20px , 20px 20px , auto , auto "
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (0.6) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "24px 24px , 24px 24px , auto , auto "
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "40px 40px , 40px 40px , auto , auto "
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, contain, cover] to [40px 40px, 40px 40px, cover, contain] at (1.5) should be [40px 40px, 40px 40px, cover, contain] assert_equals: expected "40px 40px , 40px 40px , cover , contain " but got "60px 60px , 60px 60px , auto , auto "
FAIL CSS Transitions: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (-0.3) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "0px , - 12px - 12px , 0px , - 12px - 12px "
FAIL CSS Transitions: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "0px , 0px 0px , 0px , 0px 0px "
FAIL CSS Transitions: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.3) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "0px , 12px 12px , 0px , 12px 12px "
FAIL CSS Transitions: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 20px 20px , auto 40px , 20px 20px "
FAIL CSS Transitions: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.6) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 24px 24px , auto 40px , 24px 24px "
PASS CSS Transitions: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1) should be [auto 40px, 40px 40px]
FAIL CSS Transitions: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 60px 60px , auto 40px , 60px 60px "
FAIL CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (-0.3) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "0px , - 12px - 12px , 0px , - 12px - 12px "
FAIL CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "0px , 0px 0px , 0px , 0px 0px "
FAIL CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.3) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "0px , 12px 12px , 0px , 12px 12px "
FAIL CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 20px 20px , auto 40px , 20px 20px "
FAIL CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.6) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 24px 24px , auto 40px , 24px 24px "
PASS CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1) should be [auto 40px, 40px 40px]
FAIL CSS Transitions with transition: all: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 60px 60px , auto 40px , 60px 60px "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (-0.3) should be [0px auto, 0px 0px] assert_equals: expected "0px , 0px 0px , 0px , 0px 0px " but got "0px , - 12px - 12px , 0px , - 12px - 12px "
PASS CSS Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0) should be [0px auto, 0px 0px]
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.3) should be [0px auto, 0px 0px] assert_equals: expected "0px , 0px 0px , 0px , 0px 0px " but got "0px , 12px 12px , 0px , 12px 12px "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 20px 20px , auto 40px , 20px 20px "
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.6) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 24px 24px , auto 40px , 24px 24px "
PASS CSS Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1) should be [auto 40px, 40px 40px]
FAIL CSS Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 60px 60px , auto 40px , 60px 60px "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (-0.3) should be [0px auto, 0px 0px] assert_equals: expected "0px , 0px 0px , 0px , 0px 0px " but got "0px , - 12px - 12px , 0px , - 12px - 12px "
PASS Web Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0) should be [0px auto, 0px 0px]
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.3) should be [0px auto, 0px 0px] assert_equals: expected "0px , 0px 0px , 0px , 0px 0px " but got "0px , 12px 12px , 0px , 12px 12px "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 20px 20px , auto 40px , 20px 20px "
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (0.6) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 24px 24px , auto 40px , 24px 24px "
PASS Web Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1) should be [auto 40px, 40px 40px]
FAIL Web Animations: property <background-size> from [0px auto, 0px 0px] to [auto 40px, 40px 40px] at (1.5) should be [auto 40px, 40px 40px] assert_equals: expected "auto 40px , 40px 40px , auto 40px , 40px 40px " but got "auto 40px , 60px 60px , auto 40px , 60px 60px "
FAIL CSS Transitions: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (-0.25) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 5px - 5px , - 10px - 10px , - 15px - 15px , - 25px - 25px "
PASS CSS Transitions: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px]
PASS CSS Transitions: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.25) should be [ 5px  5px, 10px 10px, 15px 15px,  25px  25px]
PASS CSS Transitions: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.5) should be [10px 10px, 20px 20px, 30px 30px,  50px  50px]
PASS CSS Transitions: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.75) should be [15px 15px, 30px 30px, 45px 45px,  75px  75px]
PASS CSS Transitions: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1) should be [20px 20px, 40px 40px, 60px 60px, 100px 100px]
PASS CSS Transitions: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1.25) should be [25px 25px, 50px 50px, 75px 75px, 125px 125px]
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (-0.25) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 5px - 5px , - 10px - 10px , - 15px - 15px , - 25px - 25px "
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.25) should be [ 5px  5px, 10px 10px, 15px 15px,  25px  25px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.5) should be [10px 10px, 20px 20px, 30px 30px,  50px  50px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.75) should be [15px 15px, 30px 30px, 45px 45px,  75px  75px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1) should be [20px 20px, 40px 40px, 60px 60px, 100px 100px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1.25) should be [25px 25px, 50px 50px, 75px 75px, 125px 125px]
FAIL CSS Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (-0.25) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 5px - 5px , - 10px - 10px , - 15px - 15px , - 25px - 25px "
PASS CSS Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px]
PASS CSS Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.25) should be [ 5px  5px, 10px 10px, 15px 15px,  25px  25px]
PASS CSS Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.5) should be [10px 10px, 20px 20px, 30px 30px,  50px  50px]
PASS CSS Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.75) should be [15px 15px, 30px 30px, 45px 45px,  75px  75px]
PASS CSS Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1) should be [20px 20px, 40px 40px, 60px 60px, 100px 100px]
PASS CSS Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1.25) should be [25px 25px, 50px 50px, 75px 75px, 125px 125px]
FAIL Web Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (-0.25) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 5px - 5px , - 10px - 10px , - 15px - 15px , - 25px - 25px "
PASS Web Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0) should be [ 0px  0px,  0px  0px,  0px  0px,   0px   0px]
PASS Web Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.25) should be [ 5px  5px, 10px 10px, 15px 15px,  25px  25px]
PASS Web Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.5) should be [10px 10px, 20px 20px, 30px 30px,  50px  50px]
PASS Web Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (0.75) should be [15px 15px, 30px 30px, 45px 45px,  75px  75px]
PASS Web Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1) should be [20px 20px, 40px 40px, 60px 60px, 100px 100px]
PASS Web Animations: property <background-size> from [0px 0px, 0px 0px, 0px 0px, 0px 0px] to [20px 20px, 40px 40px, 60px 60px, 100px 100px] at (1.25) should be [25px 25px, 50px 50px, 75px 75px, 125px 125px]
FAIL CSS Transitions: property <background-size> from [0px 0px] to [80px 80px] at (-0.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 20px - 20px , - 20px - 20px , - 20px - 20px , - 20px - 20px "
PASS CSS Transitions: property <background-size> from [0px 0px] to [80px 80px] at (0) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px]
PASS CSS Transitions: property <background-size> from [0px 0px] to [80px 80px] at (0.25) should be [ 20px  20px,  20px  20px,  20px  20px,  20px  20px]
PASS CSS Transitions: property <background-size> from [0px 0px] to [80px 80px] at (0.5) should be [ 40px  40px,  40px  40px,  40px  40px,  40px  40px]
PASS CSS Transitions: property <background-size> from [0px 0px] to [80px 80px] at (0.75) should be [ 60px  60px,  60px  60px,  60px  60px,  60px  60px]
PASS CSS Transitions: property <background-size> from [0px 0px] to [80px 80px] at (1) should be [ 80px  80px,  80px  80px,  80px  80px,  80px  80px]
PASS CSS Transitions: property <background-size> from [0px 0px] to [80px 80px] at (1.25) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px]
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px] to [80px 80px] at (-0.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 20px - 20px , - 20px - 20px , - 20px - 20px , - 20px - 20px "
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px] to [80px 80px] at (0) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px] to [80px 80px] at (0.25) should be [ 20px  20px,  20px  20px,  20px  20px,  20px  20px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px] to [80px 80px] at (0.5) should be [ 40px  40px,  40px  40px,  40px  40px,  40px  40px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px] to [80px 80px] at (0.75) should be [ 60px  60px,  60px  60px,  60px  60px,  60px  60px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px] to [80px 80px] at (1) should be [ 80px  80px,  80px  80px,  80px  80px,  80px  80px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px] to [80px 80px] at (1.25) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px]
FAIL CSS Animations: property <background-size> from [0px 0px] to [80px 80px] at (-0.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 20px - 20px , - 20px - 20px , - 20px - 20px , - 20px - 20px "
PASS CSS Animations: property <background-size> from [0px 0px] to [80px 80px] at (0) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px]
PASS CSS Animations: property <background-size> from [0px 0px] to [80px 80px] at (0.25) should be [ 20px  20px,  20px  20px,  20px  20px,  20px  20px]
PASS CSS Animations: property <background-size> from [0px 0px] to [80px 80px] at (0.5) should be [ 40px  40px,  40px  40px,  40px  40px,  40px  40px]
PASS CSS Animations: property <background-size> from [0px 0px] to [80px 80px] at (0.75) should be [ 60px  60px,  60px  60px,  60px  60px,  60px  60px]
PASS CSS Animations: property <background-size> from [0px 0px] to [80px 80px] at (1) should be [ 80px  80px,  80px  80px,  80px  80px,  80px  80px]
PASS CSS Animations: property <background-size> from [0px 0px] to [80px 80px] at (1.25) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px]
FAIL Web Animations: property <background-size> from [0px 0px] to [80px 80px] at (-0.25) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px] assert_equals: expected "0px 0px , 0px 0px , 0px 0px , 0px 0px " but got "- 20px - 20px , - 20px - 20px , - 20px - 20px , - 20px - 20px "
PASS Web Animations: property <background-size> from [0px 0px] to [80px 80px] at (0) should be [  0px   0px,   0px   0px,   0px   0px,   0px   0px]
PASS Web Animations: property <background-size> from [0px 0px] to [80px 80px] at (0.25) should be [ 20px  20px,  20px  20px,  20px  20px,  20px  20px]
PASS Web Animations: property <background-size> from [0px 0px] to [80px 80px] at (0.5) should be [ 40px  40px,  40px  40px,  40px  40px,  40px  40px]
PASS Web Animations: property <background-size> from [0px 0px] to [80px 80px] at (0.75) should be [ 60px  60px,  60px  60px,  60px  60px,  60px  60px]
PASS Web Animations: property <background-size> from [0px 0px] to [80px 80px] at (1) should be [ 80px  80px,  80px  80px,  80px  80px,  80px  80px]
PASS Web Animations: property <background-size> from [0px 0px] to [80px 80px] at (1.25) should be [100px 100px, 100px 100px, 100px 100px, 100px 100px]
FAIL CSS Transitions: property <background-size> from [0px] to [80px] at (-0.25) should be [  0px,   0px,   0px,   0px] assert_equals: expected "0px , 0px , 0px , 0px " but got "- 20px , - 20px , - 20px , - 20px "
PASS CSS Transitions: property <background-size> from [0px] to [80px] at (0) should be [  0px,   0px,   0px,   0px]
PASS CSS Transitions: property <background-size> from [0px] to [80px] at (0.25) should be [ 20px,  20px,  20px,  20px]
PASS CSS Transitions: property <background-size> from [0px] to [80px] at (0.5) should be [ 40px,  40px,  40px,  40px]
PASS CSS Transitions: property <background-size> from [0px] to [80px] at (0.75) should be [ 60px,  60px,  60px,  60px]
PASS CSS Transitions: property <background-size> from [0px] to [80px] at (1) should be [ 80px,  80px,  80px,  80px]
PASS CSS Transitions: property <background-size> from [0px] to [80px] at (1.25) should be [100px, 100px, 100px, 100px]
FAIL CSS Transitions with transition: all: property <background-size> from [0px] to [80px] at (-0.25) should be [  0px,   0px,   0px,   0px] assert_equals: expected "0px , 0px , 0px , 0px " but got "- 20px , - 20px , - 20px , - 20px "
PASS CSS Transitions with transition: all: property <background-size> from [0px] to [80px] at (0) should be [  0px,   0px,   0px,   0px]
PASS CSS Transitions with transition: all: property <background-size> from [0px] to [80px] at (0.25) should be [ 20px,  20px,  20px,  20px]
PASS CSS Transitions with transition: all: property <background-size> from [0px] to [80px] at (0.5) should be [ 40px,  40px,  40px,  40px]
PASS CSS Transitions with transition: all: property <background-size> from [0px] to [80px] at (0.75) should be [ 60px,  60px,  60px,  60px]
PASS CSS Transitions with transition: all: property <background-size> from [0px] to [80px] at (1) should be [ 80px,  80px,  80px,  80px]
PASS CSS Transitions with transition: all: property <background-size> from [0px] to [80px] at (1.25) should be [100px, 100px, 100px, 100px]
FAIL CSS Animations: property <background-size> from [0px] to [80px] at (-0.25) should be [  0px,   0px,   0px,   0px] assert_equals: expected "0px , 0px , 0px , 0px " but got "- 20px , - 20px , - 20px , - 20px "
PASS CSS Animations: property <background-size> from [0px] to [80px] at (0) should be [  0px,   0px,   0px,   0px]
PASS CSS Animations: property <background-size> from [0px] to [80px] at (0.25) should be [ 20px,  20px,  20px,  20px]
PASS CSS Animations: property <background-size> from [0px] to [80px] at (0.5) should be [ 40px,  40px,  40px,  40px]
PASS CSS Animations: property <background-size> from [0px] to [80px] at (0.75) should be [ 60px,  60px,  60px,  60px]
PASS CSS Animations: property <background-size> from [0px] to [80px] at (1) should be [ 80px,  80px,  80px,  80px]
PASS CSS Animations: property <background-size> from [0px] to [80px] at (1.25) should be [100px, 100px, 100px, 100px]
FAIL Web Animations: property <background-size> from [0px] to [80px] at (-0.25) should be [  0px,   0px,   0px,   0px] assert_equals: expected "0px , 0px , 0px , 0px " but got "- 20px , - 20px , - 20px , - 20px "
PASS Web Animations: property <background-size> from [0px] to [80px] at (0) should be [  0px,   0px,   0px,   0px]
PASS Web Animations: property <background-size> from [0px] to [80px] at (0.25) should be [ 20px,  20px,  20px,  20px]
PASS Web Animations: property <background-size> from [0px] to [80px] at (0.5) should be [ 40px,  40px,  40px,  40px]
PASS Web Animations: property <background-size> from [0px] to [80px] at (0.75) should be [ 60px,  60px,  60px,  60px]
PASS Web Animations: property <background-size> from [0px] to [80px] at (1) should be [ 80px,  80px,  80px,  80px]
PASS Web Animations: property <background-size> from [0px] to [80px] at (1.25) should be [100px, 100px, 100px, 100px]
FAIL CSS Transitions: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (-0.25) should be [ 0px  0px, 80px   0px, 0px   0px, 90px  0px] assert_equals: expected "0px 0px , 80px 0px , 0px 0px , 90px 0px " but got "- 10px - 10px , 80px - 20px , 0px - 20px , 90px - 10px "
PASS CSS Transitions: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0) should be [ 0px  0px, 80px   0px, 0px   0px, 80px  0px]
PASS CSS Transitions: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.25) should be [10px 10px, 80px  20px, 0px  20px, 70px 10px]
PASS CSS Transitions: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.5) should be [20px 20px, 80px  40px, 0px  40px, 60px 20px]
PASS CSS Transitions: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.75) should be [30px 30px, 80px  60px, 0px  60px, 50px 30px]
PASS CSS Transitions: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1) should be [40px 40px, 80px  80px, 0px  80px, 40px 40px]
PASS CSS Transitions: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1.25) should be [50px 50px, 80px 100px, 0px 100px, 30px 50px]
FAIL CSS Transitions with transition: all: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (-0.25) should be [ 0px  0px, 80px   0px, 0px   0px, 90px  0px] assert_equals: expected "0px 0px , 80px 0px , 0px 0px , 90px 0px " but got "- 10px - 10px , 80px - 20px , 0px - 20px , 90px - 10px "
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0) should be [ 0px  0px, 80px   0px, 0px   0px, 80px  0px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.25) should be [10px 10px, 80px  20px, 0px  20px, 70px 10px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.5) should be [20px 20px, 80px  40px, 0px  40px, 60px 20px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.75) should be [30px 30px, 80px  60px, 0px  60px, 50px 30px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1) should be [40px 40px, 80px  80px, 0px  80px, 40px 40px]
PASS CSS Transitions with transition: all: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1.25) should be [50px 50px, 80px 100px, 0px 100px, 30px 50px]
FAIL CSS Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (-0.25) should be [ 0px  0px, 80px   0px, 0px   0px, 90px  0px] assert_equals: expected "0px 0px , 80px 0px , 0px 0px , 90px 0px " but got "- 10px - 10px , 80px - 20px , 0px - 20px , 90px - 10px "
PASS CSS Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0) should be [ 0px  0px, 80px   0px, 0px   0px, 80px  0px]
PASS CSS Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.25) should be [10px 10px, 80px  20px, 0px  20px, 70px 10px]
PASS CSS Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.5) should be [20px 20px, 80px  40px, 0px  40px, 60px 20px]
PASS CSS Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.75) should be [30px 30px, 80px  60px, 0px  60px, 50px 30px]
PASS CSS Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1) should be [40px 40px, 80px  80px, 0px  80px, 40px 40px]
PASS CSS Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1.25) should be [50px 50px, 80px 100px, 0px 100px, 30px 50px]
FAIL Web Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (-0.25) should be [ 0px  0px, 80px   0px, 0px   0px, 90px  0px] assert_equals: expected "0px 0px , 80px 0px , 0px 0px , 90px 0px " but got "- 10px - 10px , 80px - 20px , 0px - 20px , 90px - 10px "
PASS Web Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0) should be [ 0px  0px, 80px   0px, 0px   0px, 80px  0px]
PASS Web Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.25) should be [10px 10px, 80px  20px, 0px  20px, 70px 10px]
PASS Web Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.5) should be [20px 20px, 80px  40px, 0px  40px, 60px 20px]
PASS Web Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (0.75) should be [30px 30px, 80px  60px, 0px  60px, 50px 30px]
PASS Web Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1) should be [40px 40px, 80px  80px, 0px  80px, 40px 40px]
PASS Web Animations: property <background-size> from [0px 0px, 80px 0px] to [40px 40px, 80px 80px, 0px 80px] at (1.25) should be [50px 50px, 80px 100px, 0px 100px, 30px 50px]
